fix: 학생 전시 페이지와 동일한 구조로 졸업 전시 페이지 구조 수정 (#103)
