+++ apps/client/public/data/gallery.json
@@ -1,5 +1,5 @@
 {
-  "year": "2024",
+  "year": 2024,
   "categoryList": [
     "All",
     "Smart Mobility",
