+++ apps/client/src/components/Gallery/Exhibition.types.ts
