+++ apps/client/src/components/Gallery/GraduationGallery/GraduationExhibitionGallery.styled.ts
