+++ apps/client/src/components/Gallery/GraduationGallery/Piece/Piece.styled.ts
