+++ apps/client/src/components/Gallery/GraduationGallery/Piece/Piece.tsx
