+++ apps/client/src/components/Gallery/StudentGallery/StudentExhibitionGallery.styled.ts
