+++ apps/client/src/components/Gallery/StudentGallery/StudentExhibitionGallery.tsx
